llmops-api/requirements.txt
alembic==1.13.3
Flask==3.0.3
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
injector==0.22.0
openai==1.51.0
pytest==8.3.3
python-dotenv==1.0.1
WTForms==3.1.2
